Refactor into JobSample
